+++ src/main/java/roomescape/presentation/ViewController.java
@@ -4,10 +4,20 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class HomeController {
+public class ViewController {
 
     @GetMapping("/")
     public String home() {
         return "home";
+    }
+
+    @GetMapping("/reservation")
+    public String getReservation() {
+        return "new-reservation";
+    }
+
+    @GetMapping("/time")
+    public String getReservationTime() {
+        return "time";
     }
 }
